remove Ukranian translation (uk)
This is to stay within the weblate "basic" plan (15 languages), for now.

In the future may request a libre plan or exception; we use pretty few
strings for this project so far. But Ukranian was requested on weblate
but then never contributed to, so removing seems like an easy short-term
resolution.

Czech and Greek also were requested but have never had contributions.