change health check from .exists(index) to .mapping(index)
In cases where the cluser leader node is unavilable, the health check
was returning false even when the local node had full shard replicas and
could return requests.

A refinement of this change would be to use the /<index>/_count API
endpoint to ensure that the "failed" and "skipped" shard numbers are 0
(aka, "successful == total"). However, not sure where that endpoint is
exposed in the elasticsearch-py API. the CatClient method doesn't seem
right.