web: access_redirect_fallback mechanism
This adds a helper code path that "tries harder" to find an access link,
by querying the fatcat API directly to look for any file from any
release associated with the work. If it finds a match, it does the
redirect as usual (but does log the incident).

If no match can be found, there is now a more helpful access-specific
404 error page.

If the *work* is a 404, the generic error page is shown.