add docker file for local development
I'm no docker wizard so it's possible there are better ways of going
about any of this but this is working well locally on Debian bookworm.

I did this because I was struggling to get the Python dependencies to
install. I think even with pyenv/pipenv dependencies that required
native compilation were failing. Instead of embarking on updating all
the dependencies I chose to just get things running on focal in a docker
container.