port fatcat.wiki website
This is a large commit that brings almost the entirety of the fatcat web
interface into this codebase.

The things that *aren't* ported:

- authenticated features (including editing)
- the guide
- sitemap route (we would like to get to this in the future, though)

The things that are changed:

- the about page
- the "pseudo-api" is tweaked in minor ways

I strove to keep all of the paths identical so we can have 1:1
redirection from fatcat.wiki when the time comes.

I have done some QA but this work needs plenty more; I have also not
begun porting over any tests yet.