Add new read-only, shim API
This slightly worrisome commit creates a new FastAPI api for fatcat.
It's a small subset of the v0 API -- just the most useful read-only
routes.

It's implemented as a transparent pass through to the rust-based fatcatd
via the code-generated python API client.

This is a kludge; however, it is a forward looking kludge.

I'm merging this in the short term so we can advertise an API for
Force11.

In the medium term, the intention is to reimplement the API endpoints in
pure python and deprecate them from the rust fatcatd. The approach taken
in this commit allows that to be done incrementally.

In the long term, we'll need to decide where and how to house the
_write_ endpoints of the fatcat API which we are moving to use only
internally but still need to live somewhere.

NB! The generated docs for this API are not good because I had to fall
back to exposing our "types" as dictionaries. This is because we don't
have fatcat datatypes described in PyDantic yet. I did attempt to do
that but it's a ton of work I can't get done before the conference.